Adiconado exceptions personalizadas, falta corrigir o and point de collaborators

+++ src/main/java/com/logistica/order/OrderService.java
@@ -1,8 +1,7 @@
 package com.logistica.order;
 
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
+import com.logistica.exception.ResourceNotFoundException;
 
 import java.util.List;
 
@@ -18,9 +17,6 @@ public class OrderService {
     // Salvar Ordem
     public OrderDTO saveOrder(OrderDTO orderDTO){
         OrderModel order = new OrderModel(orderDTO.product(), orderDTO.packaging(), orderDTO.recipient());
-        order.setProduct(orderDTO.product());
-        order.setPackaging(orderDTO.packaging());
-        order.setRecipient(orderDTO.recipient());
         OrderModel savedOrder = repository.save(order);
         return new OrderDTO(savedOrder.getId(), savedOrder.getOrderNumber(), savedOrder.getProduct(), savedOrder.getPackaging(), savedOrder.getRecipient());
     }
@@ -42,14 +38,15 @@ public class OrderService {
     // Buscar Ordem por ID
     public OrderDTO getOrderById(Long id){
         OrderModel order = repository.findById(id)
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Order not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("Pedido não encontrado"));
         return new OrderDTO(order.getId(), order.getOrderNumber(), order.getProduct(), order.getPackaging(), order.getRecipient());
     }
 
     //Atualizar Ordem
     public OrderDTO updateOrder(Long id, OrderDTO orderDTO){
         OrderModel order = repository.findById(id)
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Order not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("Pedido não encontrado"));
+        
         // Apenas atualiza o número do pedido se um novo valor for fornecido
         if (orderDTO.orderNumber() != null) {
             order.setOrderNumber(orderDTO.orderNumber());
@@ -63,6 +60,10 @@ public class OrderService {
 
     // Deletar Ordem
     public void deleteOrder(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Pedido não encontrado");
+        }
         repository.deleteById(id);
     }
 }
+// End of OrderService
